requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
mysql-connector-python==8.0.31
stripe==5.2.0
python-dotenv==0.21.0
requests==2.28.2
pika==1.3.1